run_make_support: move `env_var{,_os}` into `env_checked` module
